Switch to okhttp 3 for HTTP and WS
Switch to okhttp3,  removed okhttp 2.7, switch WS library to okhttp3-ws.

#231 #249 #251 #217 #177 #138

+++ src/test/java/com/ibm/watson/developer_cloud/language_translation/v2/LanguageTranslationIT.java
@@ -133,7 +133,7 @@ public class LanguageTranslationIT extends WatsonServiceTest {
 
   private void testTranslationResult(String text, String result, TranslationResult translationResult) {
     assertNotNull(translationResult);
-    assertEquals(translationResult.getWordCount(), text.split(" ").length);
+    assertEquals(translationResult.getWordCount().intValue(), text.split(" ").length);
     assertNotNull(translationResult.getTranslations());
     assertNotNull(translationResult.getTranslations().get(0).getTranslation());
     assertEquals(result, translationResult.getTranslations().get(0).getTranslation());
